chore: resolve linting errors
